move image scanning methods to gated file
