+++ pkg/util/trivy/trivy.go
@@ -22,7 +22,6 @@ import (
 	"github.com/aquasecurity/trivy/pkg/fanal/analyzer"
 	"github.com/aquasecurity/trivy/pkg/fanal/applier"
 	"github.com/aquasecurity/trivy/pkg/fanal/artifact"
-	image2 "github.com/aquasecurity/trivy/pkg/fanal/artifact/image"
 	local2 "github.com/aquasecurity/trivy/pkg/fanal/artifact/local"
 	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
 	"github.com/aquasecurity/trivy/pkg/fanal/walker"
@@ -304,18 +303,6 @@ func (c *Collector) ScanFilesystem(ctx context.Context, path string, scanOptions
 	return c.buildReport(trivyReport, cache.lastBlobID), nil
 }
 
-func (c *Collector) fixupCacheKeyForImgMeta(ctx context.Context, artifact artifact.Artifact, imgMeta *workloadmeta.ContainerImageMetadata, cache CacheWithCleaner) error {
-	// The artifact reference is only needed to clean up the blobs after the scan.
-	// It is re-generated from cached partial results during the scan.
-	artifactReference, err := artifact.Inspect(ctx)
-	if err != nil {
-		return err
-	}
-
-	cache.setKeysForEntity(imgMeta.EntityID.ID, append(artifactReference.BlobIDs, artifactReference.ID))
-	return nil
-}
-
 func (c *Collector) scan(ctx context.Context, artifact artifact.Artifact, applier applier.Applier) (*types.Report, error) {
 	localScanner := local.NewScanner(applier, c.osScanner, c.langScanner, c.vulnClient)
 	s := scanner.NewScanner(localScanner, artifact)
@@ -331,29 +318,6 @@ func (c *Collector) scan(ctx context.Context, artifact artifact.Artifact, applie
 	}
 
 	return &trivyReport, nil
-}
-
-func (c *Collector) scanImage(ctx context.Context, fanalImage ftypes.Image, imgMeta *workloadmeta.ContainerImageMetadata, scanOptions sbom.ScanOptions) (sbom.Report, error) {
-	cache, err := c.getCache()
-	if err != nil {
-		return nil, err
-	}
-
-	imageArtifact, err := image2.NewArtifact(fanalImage, cache, getDefaultArtifactOption(scanOptions))
-	if err != nil {
-		return nil, fmt.Errorf("unable to create artifact from image, err: %w", err)
-	}
-
-	if err := c.fixupCacheKeyForImgMeta(ctx, imageArtifact, imgMeta, cache); err != nil {
-		return nil, fmt.Errorf("unable to fixup cache key for image, err: %w", err)
-	}
-
-	trivyReport, err := c.scan(ctx, imageArtifact, applier.NewApplier(cache))
-	if err != nil {
-		return nil, fmt.Errorf("unable to marshal report to sbom format, err: %w", err)
-	}
-
-	return c.buildReport(trivyReport, trivyReport.Metadata.ImageID), nil
 }
 
 func (c *Collector) buildReport(trivyReport *types.Report, id string) *Report {
